refactor: add app.rs
