+++ frontend/trainee/src/lib.rs
@@ -2,8 +2,9 @@ mod app;
 
 use wasm_bindgen::prelude::*;
 use yew::prelude::*;
+use crate::app::app as app_root;
 
 #[wasm_bindgen(start)]
 pub fn run_app() {
-    App::<app::App>::new().mount_to_body();
+    App::<app_root::App>::new().mount_to_body();
 }
